[feat][accounts] Rename users package to accounts
