+++ accounts/src/main/kotlin/pl/edu/pw/ia/accounts/application/converter/AccountCommandConverters.kt
@@ -1,0 +1,18 @@
+package pl.edu.pw.ia.accounts.application.converter
+
+import java.util.UUID
+import org.springframework.core.convert.converter.Converter
+import pl.edu.pw.ia.shared.config.TypeConverter
+import pl.edu.pw.ia.shared.domain.command.CreateAccountCommand
+import pl.edu.pw.ia.accounts.application.model.CreateAccountRequest
+
+@TypeConverter
+class CreateAccountRequestToCreateAccountCommandConverter : Converter<CreateAccountRequest, CreateAccountCommand> {
+
+	override fun convert(source: CreateAccountRequest): CreateAccountCommand? =
+		CreateAccountCommand(
+			accountId = UUID.randomUUID(),
+			name = source.name,
+			email = source.email,
+		)
+}
