+++ accounts/src/main/kotlin/pl/edu/pw/ia/accounts/domain/query/repository/AccountViewRepository.kt
@@ -1,0 +1,11 @@
+package pl.edu.pw.ia.accounts.domain.query.repository
+
+import java.util.UUID
+import pl.edu.pw.ia.accounts.domain.query.view.AccountView
+
+interface AccountViewRepository {
+
+	fun save(user: AccountView)
+
+	fun findById(id: UUID): AccountView?
+}
