+++ accounts/src/main/kotlin/pl/edu/pw/ia/accounts/infrastructure/query/repository/database/SpringAccountViewEntityRepository.kt
@@ -1,0 +1,8 @@
+package pl.edu.pw.ia.accounts.infrastructure.query.repository.database
+
+import org.springframework.data.mongodb.repository.ReactiveMongoRepository
+import org.springframework.stereotype.Repository
+import pl.edu.pw.ia.accounts.infrastructure.query.entity.AccountViewEntity
+
+@Repository
+interface SpringAccountViewEntityRepository : ReactiveMongoRepository<AccountViewEntity, String>
