+++ accounts/src/test/kotlin/pl/edu/pw/ia/accounts/AccountsApplicationTests.kt
@@ -1,0 +1,12 @@
+package pl.edu.pw.ia.accounts
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest(classes = [TestAccountsApplication::class])
+class AccountsApplicationTests {
+
+	@Test
+	fun contextLoads() {
+	}
+}
