--- accounts/src/test/kotlin/pl/edu/pw/ia/users/AccountsApplicationTests.kt
@@ -1,12 +1,0 @@
-package pl.edu.pw.ia.users
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-
-@SpringBootTest(classes = [TestAccountsApplication::class])
-class AccountsApplicationTests {
-
-	@Test
-	fun contextLoads() {
-	}
-}
