--- accounts/src/test/kotlin/pl/edu/pw/ia/users/TestAccountsApplication.kt
@@ -1,25 +1,0 @@
-package pl.edu.pw.ia.users
-
-import org.springframework.boot.fromApplication
-import org.springframework.boot.test.context.TestConfiguration
-import org.springframework.boot.testcontainers.service.connection.ServiceConnection
-import org.springframework.boot.with
-import org.springframework.context.annotation.Bean
-import org.testcontainers.containers.MongoDBContainer
-import org.testcontainers.utility.DockerImageName
-import pl.edu.pw.ia.shared.SharedApplication
-
-@TestConfiguration(proxyBeanMethods = false)
-class TestAccountsApplication {
-
-	@Bean
-	@ServiceConnection
-	fun mongoDbContainer(): MongoDBContainer {
-		return MongoDBContainer(DockerImageName.parse("mongo:latest"))
-	}
-
-}
-
-fun main(args: Array<String>) {
-	fromApplication<SharedApplication>().with(TestAccountsApplication::class).run(*args)
-}
